requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cycler==0.10.0
defusedxml==0.5.0
Django==2.1.2
django-braces==1.13.0
django-filter==2.0.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.8.2
dnspython==1.15.0
eventlet==0.24.1
gevent==1.3.6
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
kiwisolver==1.0.1
lxml==4.2.5
Markdown==3.0.1
matplotlib==3.0.0
monotonic==1.5
numpy==1.15.2
oauthlib==2.1.0
pandas==0.23.4
psutil==5.4.7
psycopg2==2.7.5
pycparser==2.19
PyJWT==1.6.4
pyparsing==2.2.2
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
SQLAlchemy==1.2.12
urllib3==1.23
wincertstore==0.2
uwsgi
mysqlclient
django-redis==4.9.0
